[fix] fix async issue with getPendingOrders test
